revno: 2123 message:   update design data entry form function(patient)
committer: Tri <Tri@Tri-Laptop>
branch nick: dhis2
timestamp: Sat 2010-11-20 20:17:20 +0700

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-patient/src/main/java/org/hisp/dhis/patient/action/dataentryform/DelDataEntryFormAction.java
@@ -105,7 +105,11 @@ public class DelDataEntryFormAction
 
         programStageService.updateProgramStage( programStage );
 
-        dataEntryFormService.deleteDataEntryForm( dataEntryForm );
+        if ( programStageService.getAllProgramStageByDataEntryForm( dataEntryForm ).isEmpty() )
+        {
+
+            dataEntryFormService.deleteDataEntryForm( dataEntryForm );
+        }
 
         return SUCCESS;
     }
